add lib for m.analyse.buildings
